+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Reduced OpenGL requirements from 3.3 to 2.1 for desktop.
 * Added copy marker text to clipboard  #298
 * Added optional show time for vertical x-axis markers in Waveform widget #297
+* Fixed Waveform summary to respect min/max lines setting.
 * Updated build
   * pyjoulescope_driver from 1.7.2 to 1.7.4, fixes #293
   * Python 3.11 to 3.12